LayoutTests/platform/mac/imported/w3c/web-platform-tests/media-source/SourceBuffer-abort-updating-expected.txt


FAIL SourceBuffer#abort() (video/webm; codecs="vorbis,vp8") : Check the algorithm when the updating attribute is true. assert_unreached: Browser doesn't support the MIME used in this test: video/webm; codecs="vorbis,vp8" Reached unreachable code
PASS SourceBuffer#abort() (video/mp4) : Check the algorithm when the updating attribute is true.
